Fix installer script
Once we merged the pedestal-restructure branch to master EasyEngine started not liking the fact that the `public/` directory was filled with the theme files. Boo.

Instead we leave `public/` empty, let EasyEngine install WordPress, clone the Pedestal repo to a tmp directory, move the `.git/` directory to `public/`, force a `git reset` to magically weave our repo files with the installed WordPress files.